+++ src/geometry/PTSceneNode.cxx
@@ -99,13 +99,13 @@ GLfloat			PhotonTorpedoSceneNode::PTRenderNode::core[9][2];
 GLfloat			PhotonTorpedoSceneNode::PTRenderNode::corona[8][2];
 const GLfloat		PhotonTorpedoSceneNode::PTRenderNode::ring[8][2] = {
 				{ 1.0f, 0.0f },
-				{ M_SQRT1_2, M_SQRT1_2 },
+				{ (float)M_SQRT1_2, (float)M_SQRT1_2 },
 				{ 0.0f, 1.0f },
-				{ -M_SQRT1_2, M_SQRT1_2 },
+				{ (float)-M_SQRT1_2, (float)M_SQRT1_2 },
 				{ -1.0f, 0.0f },
-				{ -M_SQRT1_2, -M_SQRT1_2 },
+				{ (float)-M_SQRT1_2, (float)-M_SQRT1_2 },
 				{ 0.0f, -1.0f },
-				{ M_SQRT1_2, -M_SQRT1_2 }
+				{ (float)M_SQRT1_2, (float)-M_SQRT1_2 }
 			};
 
 PhotonTorpedoSceneNode::PTRenderNode::PTRenderNode(
@@ -149,9 +149,9 @@ void			PhotonTorpedoSceneNode::PTRenderNode::render()
     if (!RENDERER.isSameFrame()) {
       numFlares = 3 + int(3.0f * (float)bzfrand());
       for (int i = 0; i < numFlares; i++) {
-	theta[i] = 2.0f * M_PI * (float)bzfrand();
-	phi[i] = (float)bzfrand() - 0.5f;
-	phi[i] *= 2.0f * M_PI * fabsf(phi[i]);
+	theta[i] = (float)(2.0 * M_PI * bzfrand());
+	phi[i] = (float)(bzfrand() - 0.5);
+	phi[i] *= (float)(2.0 * M_PI * fabsf(phi[i]));
       }
     }
